test(research): widen expected quantity before multiplication
Continue PR107 from actual remote 75d54e7 without rewriting history. The bounded constant-price fixture has side in {-1,1} and i in [1,4096], so the old intermediate product was representable; make its intended int64 quantity arithmetic explicit before multiplication rather than depending on range inference.

One test line only. Preserve all 4096 iterations per side, original assertions, production code, SDK interfaces, CMake graph, workflow/security configuration and required checks. No CodeQL suppression, scan dismissal, permission change or success status is manufactured.

The complete tree matches the local staged tree c66f8b27f4caf1bc1172ae5b1ac831ecd51970e0. The actual analytics executable passes GCC strict-warning Release and Clang ASan/UBSan locally. Full exact-head remote CI and security scan must complete anew; 75d54e7 artifacts are historical evidence, not acceptance of this commit. All compatibility retention and archival limits remain.

+++ tests/research/analytics_tests.cpp
@@ -60,7 +60,7 @@ void BoundedCostAndAtomicRejection() {
             const auto fill = F("constant-" + std::to_string(i), i, side, 1, max);
             Check(ledger.Apply(fill) && !ledger.Apply(fill), "constant fill identity");
             const auto account = ledger.Mark(max);
-            Check(account.quantity == side * i && account.averageEntry == max &&
+            Check(account.quantity == static_cast<std::int64_t>(side) * i && account.averageEntry == max &&
                   account.unrealized == 0 && account.realizedGross == 0 && account.equity == 1000,
                   "constant-price cost and equity invariants");
         }
